+++ lib/cocoapods-core/gem_version.rb
@@ -1,5 +1,5 @@
 module Pod
   # The version of the cocoapods-core.
   #
-  CORE_VERSION = '1.6.0.beta.1'.freeze unless defined? Pod::CORE_VERSION
+  CORE_VERSION = '1.6.0.beta.2'.freeze unless defined? Pod::CORE_VERSION
 end